fix widget generator

+++ lib/tasks/init_content.rb
@@ -137,11 +137,11 @@ module WidgetLib
 
     end
 
-    def combine_css
+    def combine_css(force=true)
       widget = WidgetLib::Generate.new
       puts '--- Combine CSS'
       combined = "#{widget.css_path}/combined.css"
-      File.delete(combined) if File.exist? combined
+      File.delete(combined) if File.exist? combined if force
       system("cat #{widget.css_path}/widgets/*.css > #{combined}")
       puts '--- Delete uncombined CSS'
       FileUtils.rm_rf("#{widget.css_path}/widgets")
